Upgrade IdentifiableFilter
Add priority
Extract resource and resource pack name on construction
Make logging consistent
